basic styles, file clean up
