+++ client/BookApp-React/src/pages/Author/ViewAuthors.jsx
@@ -26,10 +26,11 @@ export default function ViewAuthors() {
       //Eduardo map throught the books state. And display the data. Could do it here or make a new component that takes in props
       return (
         <div className="authorcontainer">
+          <h1 className='heading'>Author's</h1>
           {authors.map((item) => (
            
             <div className="authorsmain">
-              <Link to= {`/viewauthors/${item.id}`}>
+              <Link className="authors" to= {`/viewauthors/${item.id}`}>
                 <h1>{item.name}</h1>
               </Link>
             </div>
